pyproject.toml
[tool.poetry]
name = "chat2sqlgraph"
version = "0.1.0"
description = ""
authors = ["CYY <1597081640@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pyyaml = "^6.0.2"
psycopg2 = "^2.9.9"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
pydantic = "^2.9.2"
pymysql = {extras = ["rsa"], version = "^1.1.1"}
rsa = "^4.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
